Implement image upload feature.
The image upload feature replaces the the hotlinking of images from the web. This way, users can upload image right from  their devices.

A side effect of this implementation is that, the 'price' field in the 'new campground is no longer there. A minor issue that I will resolve shortly.